# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: Data Service
  description: Service to expose StackingUp DataDB.
  contact:
    name: StackingUp
    email: stackingup13@gmail.com
    url: 'https://stacking-up.github.io'
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
paths:
...
  /api/v1/spaces:
    get:
      tags:
        - Spaces
      x-swagger-router-controller: SpaceController
      description: Returns all spaces
      operationId: getSpaces
      parameters: 
        - name: offset
          in: query
          description: 'Number of elements to skip for pagination'
          required: false
          schema:
            type: integer
            format: int32

        - name: limit
          in: query
          description: 'Number of elements for pagination'
          required: false
          schema:
            type: integer
            format: int32

        - name: tag
          in: query
          description: 'Filter for given tags'
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              enum: 
                - FLOOR_1
                - FLOOR_2
                - FLOOR_3UP
                - HOUSE_ROOM
                - FLAT_ROOM
                - BASEMENT
                - GARAGE
                - STORAGE_ROOM
                - INDUSTRIAL_WAREHOUSE
                - OFFICE_ROOM
                - PENTHOUSE
                - ELEVATOR
                - WET
                - DRY
                - COLD
                - WARM
                - HOT
                - SECURITY_ALARM
                - VIDEO_MONITORING
                - FIRE_ALARM
                - GROUND_FLOOR
                - SOCKET
                - INDOOR
                - OUTDOOR
                - NARROW_ACCESS
                - MEDIUM_WIDTH_ACCESS
                - WIDE_ACCESS
                - OTHERS
                
        - name: shared
          in: query
          description: 'Filter for shared category'
          required: false
          schema:
            type: boolean

        - name: minDim
          in: query
          description: 'Min value included for dimension filtering (Optional)'
          required: false
          schema:
            type: string

        - name: maxDim
          in: query
          description: 'Max value included for dimension filtering (Optional)'
          required: false
          schema:
            type: string

        - name: minPriceHour
          in: query
          description: 'Min value included for price/hour filtering (Optional)'
          required: false
          schema:
            type: string

        - name: maxPriceHour
          in: query
          description: 'Max value included for price/hour filtering (Optional)'
          required: false
          schema:
            type: string

        - name: minPriceDay
          in: query
          description: 'Min value included for price/day filtering (Optional)'
          required: false
          schema:
            type: string

        - name: maxPriceDay
          in: query
          description: 'Max value included for price/day filtering (Optional)'
          required: false
          schema:
            type: string

        - name: minPriceMonth
          in: query
          description: 'Min value included for price/month filtering (Optional)'
          required: false
          schema:
            type: string

        - name: maxPriceMonth
          in: query
          description: 'Max value included for price/month filtering (Optional)'
          required: false
          schema:
            type: string
        
        - name: isRentPerHour
          in: query
          description: 'Check if a space can be rented per hours'
          required: false
          schema:
            type: boolean

        - name: isRentPerDay
          in: query
          description: 'Check if a space can be rented per days'
          required: false
          schema:
            type: boolean
        
        - name: isRentPerMonth
          in: query
          description: 'Check if a space can be rented per months'
          required: false
          schema:
            type: boolean

        - name: orderBy
          in: query
          description: 'Order by price or initial date'
          required: false
          schema:
            type: string

        - name: search
          in: query
          description: 'Search the given text in title, description and location'
          required: false
          schema:
            type: string

        
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/space'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      security:
        - cookieAuth: []
      tags:
        - Spaces
      x-swagger-router-controller: SpaceController
      description: Checks the validity of a logged-in user's token and publishes the space
      operationId: postSpace      
      requestBody:
        x-name: body
        description: Space object to be published
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/space'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
...
  /api/v1/smartSearch/spaces:
    post:
      tags:
        - SmartSearch
      x-swagger-router-controller: SmartSearchController
      description: Search for spaces through an advanced algorithm
      operationId: getSpaces
      parameters: 
        - name: latitude
          in: query
          description: 'User latitude'
          required: false
          schema:
            type: number
            format: float
        
        - name: longitude
          in: query
          description: 'User longitude'
          required: false
          schema:
            type: number
            format: float

      requestBody:
        x-name: items
        description: Items to search for
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/item'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/smartSearchResult'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/smartSearch/spaces/train:
    post:
      tags:
        - SmartSearch
      x-swagger-router-controller: SmartSearchController
      description: Search for spaces through an advanced algorithm
      operationId: postTrainSpaces
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /api/v1/smartSearch/renters:
    get:
      tags:
        - SmartSearch
      x-swagger-router-controller: SmartSearchController
      description: Search for renters through an advanced algorithm
      operationId: getRenters
      parameters: 
        - name: spaceId
          in: query
          description: 'space to search renters for'
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/smartSearchResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
...
components:
  schemas:
    errorModel:
      type: string
...
    item:
      type: object
      required:
        - type
        - dimensions
      properties:
        type:
          type: string
        dimensions:
          type: string
...
    space:
      type: object
      required:
        - name
        - description
        - initialDate
        - location
        - dimensions
        - shared
        - ownerId
      properties:
        name:
          type: string
        description:
          type: string
        initialDate:
          type: string
          format: date-time
        finalDate:
          type: string
          format: date-time
        location:
          type: string
        dimensions:
          type: string
        priceHour:
          type: number
          format: double
        priceDay:
          type: number
          format: double
        priceMonth:
          type: number
          format: double
        shared:
          type: boolean
        ownerId:
          type: integer
          format: int32
        tags:
          type: array
          items:
            type: string
        image:
          type: array
          items:
            type: string
            format: byte
...
    smartSearchResult:
      type: object
      required:
        - id
        - score
      properties:
        id:
          type: integer
          format: int32
        score:
          type: number
          format: double
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: authToken